all: Move remainder of graph::data::schema into graph::schema
