+++ graph/src/data/mod.rs
@@ -4,9 +4,6 @@ pub mod subgraph;
 /// Data types for dealing with GraphQL queries.
 pub mod query;
 
-/// Data types for dealing with GraphQL schemas.
-pub mod schema;
-
 /// Data types for dealing with storing entities.
 pub mod store;
