+++ graph/src/schema/input_schema.rs
@@ -6,14 +6,13 @@ use anyhow::{anyhow, Error};
 use crate::components::store::{EntityKey, EntityType, LoadRelatedRequest};
 use crate::data::graphql::ext::DirectiveFinder;
 use crate::data::graphql::{DirectiveExt, DocumentExt, ObjectTypeExt, TypeExt, ValueExt};
-use crate::data::schema::{Schema, SchemaValidationError};
 use crate::data::store::{self, scalar};
 use crate::prelude::q::Value;
 use crate::prelude::{s, DeploymentHash};
 use crate::schema::api_schema;
 
 use super::fulltext::FulltextDefinition;
-use super::ApiSchema;
+use super::{ApiSchema, Schema, SchemaValidationError};
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct InputSchema {
