+++ graphql/src/store/resolver.rs
@@ -2,16 +2,13 @@ use std::collections::BTreeMap;
 use std::result;
 use std::sync::Arc;
 
+use graph::components::store::*;
+use graph::data::graphql::{object, ObjectOrInterface};
 use graph::data::query::Trace;
 use graph::data::value::Object;
-use graph::data::{
-    graphql::{object, ObjectOrInterface},
-    schema::META_FIELD_TYPE,
-};
 use graph::prelude::*;
-use graph::schema::ErrorPolicy;
-use graph::schema::{ast as sast, ApiSchema};
-use graph::{components::store::*, data::schema::BLOCK_FIELD_TYPE};
+use graph::schema::{ast as sast, ApiSchema, META_FIELD_TYPE};
+use graph::schema::{ErrorPolicy, BLOCK_FIELD_TYPE};
 
 use crate::execution::ast as a;
 use crate::metrics::GraphQLMetrics;
