+++ server/index-node/src/schema.rs
@@ -1,4 +1,7 @@
-use graph::{prelude::*, schema::ApiSchema};
+use graph::{
+    prelude::*,
+    schema::{ApiSchema, Schema},
+};
 
 lazy_static! {
     pub static ref SCHEMA: Arc<ApiSchema> = {
